añadido test pipeClient
